requirements.txt
gradio
dashscope
openai
pillow
requests
